// frontend/src/app/core/generated/circabc/model/keywordDefinition.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface KeywordDefinition { 
    readonly id?: string;
    name?: string;
    /**
     * The object that is used to compile all the translations of a node into a JSON object. It is basically composed of a map with a key languaguage code and its value 
     */
    title: { [key: string]: string; };
}